refactor: moved code into smaller files
